Fp 145 develop order api endpoint (#5)
update order processing

+++ src/main/java/ua/flowerista/shop/services/BouqueteService.java
@@ -1,22 +1,12 @@
 package ua.flowerista.shop.services;
 
-import java.io.IOException;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
-import jakarta.transaction.Transactional;
 import ua.flowerista.shop.dto.BouqueteCardDto;
 import ua.flowerista.shop.dto.BouqueteDto;
 import ua.flowerista.shop.dto.BouqueteSmallDto;
@@ -26,6 +16,11 @@ import ua.flowerista.shop.models.Bouquete;
 import ua.flowerista.shop.models.BouqueteSize;
 import ua.flowerista.shop.models.Size;
 import ua.flowerista.shop.repo.BouqueteRepository;
+
+import java.io.IOException;
+import java.math.BigInteger;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -57,7 +52,7 @@ public class BouqueteService {
 		dto.setImageUrls(imageUrls);
 		repo.save(mapper.toEntity(dto));
 	}
-	
+
 	public void insert(BouqueteDto dto) {
 		repo.save(mapper.toEntity(dto));
 	}
@@ -87,43 +82,43 @@ public class BouqueteService {
 		return repo.findTop5ByOrderByDiscountDesc().stream().map(bouquete -> mapper.toSmallDto(bouquete))
 				.collect(Collectors.toList());
 	}
-	
+
 	public Page<BouqueteSmallDto> getBouquetesCatalogFiltered (List<Integer> flowerIds, List<Integer> colorIds, Integer minPrice, Integer maxPrice, Boolean sortByNewest,Boolean sortByPriceHighToLow,Boolean sortByPriceLowToHigh, int page) {
 		Pageable pageable = PageRequest.of(page - 1, 20);
 		return repo.findByFilters(flowerIds, colorIds, minPrice, maxPrice, sortByNewest, sortByPriceHighToLow, sortByPriceLowToHigh, pageable).map(mapper::toSmallDto);
-		
+
 	}
-	
+
     public PriceRangeDto getMinMaxPrices() {
         Integer minPrice = repo.findMinPrice();
         Integer maxPrice = repo.findMaxPrice();
 
         return new PriceRangeDto(minPrice, maxPrice);
     }
-    
+
     public BouqueteCardDto getById (int id) {
     	return mapper.toCardDto(repo.getReferenceById(id));
     }
-    
+
     public List<BouqueteSmallDto> searchBouquetesByName(String name) {
         if (name == null || name.length() < 3) {
             return Collections.emptyList();
         }
         return repo.searchByName(name).stream().map(boquete -> mapper.toSmallDto(boquete)).collect(Collectors.toList());
     }
-    
+
     public void test () {
     	Bouquete bouquete = new Bouquete();
     	BouqueteSize sizeS = new BouqueteSize();
     	sizeS.setSize(Size.SMALL);
-    	sizeS.setDefaultPrice(123);
+    	sizeS.setDefaultPrice(BigInteger.valueOf(123));
     	BouqueteSize sizeM = new BouqueteSize();
     	sizeM.setSize(Size.MEDIUM);
-    	sizeM.setDefaultPrice(140);
+    	sizeM.setDefaultPrice(BigInteger.valueOf(140));
     	BouqueteSize sizeL = new BouqueteSize();
     	sizeL.setSize(Size.LARGE);
-    	sizeL.setDefaultPrice(180);
-    	
+    	sizeL.setDefaultPrice(BigInteger.valueOf(180));
+
     	Set<BouqueteSize> sizes = new HashSet<>();
     	sizes.add(sizeS);
     	sizes.add(sizeM);
